Configuração de Handler para CreateArquivoFromUpload
Foi criado um handler completo contendo variável de Request, Response, Validator para CreateArquivoFromUpload.

-> Instalação de FluentValidation para validar no backend e ter um reforço de validação.

-> Abertura para leitura de xml dos comentários postos a cima dos métodos do controller, podento ter uma descrição de tudo no Swagger.

-> Uso do Repository + EF com classe Base para facilitar consultar/ações comuns.

-> Adição de AutoMapper com classe global sem precisar ficar criando CreateMap.

-> Criação de Resources para mensagens de validações tanto Api quanto Application Handler.

-> Adição de AllowedOrigins para liberação de Cors (Browsers que acessam a Api), foi configurado as rotas de acesso permitido no appsettings.

-> Adição de ApiKey e ApiSecret para acesso a Api fora rotas habilitadas via AllowedOrigins.

-> Configuração do Swagger.

+++ backend/CaseTecnico.MRA.Domain/Entities/Arquivo.cs
@@ -3,7 +3,7 @@ namespace CaseTecnico.MRA.Domain.Entities;
 
 public sealed class Arquivo : BaseEntity
 {
-    public int Estabelecimento { get; set; }
+    public long Estabelecimento { get; set; }
 
     public DateTime DataProcessamento { get; set; }
 
@@ -11,7 +11,7 @@ public sealed class Arquivo : BaseEntity
 
     public DateTime? PeriodoFinal { get; set; }
 
-    public int Sequencia { get; set; }
+    public long Sequencia { get; set; }
 
     public string EstruturaImportada { get; set; } = null!;
